Fixed an error in the LeNet challenge instructions.
Step 7 in the instructions for the LeNet challenge suggests that
the third convolution operation (conv_3) should contain 125 filters
while the comment in the code block indicates it should have 120
filters. Adding 125 filters will work, producing comparable
results/model performance. However, the subsequent visualisation
of the activation or feature maps for layer3 will fail when the
subplot index exceeds 120.

This change simply corrects the challenge instructions to be
consistent with the code comments and subsequent visualisation
to avoid confusion.